refactor: merge hooks and atoms for receiver and transmitter
