--- src/components/DistributedReversi/recoil.transmit.mts
@@ -1,40 +1,0 @@
-import { writer } from '../../util/recoil/selector.mts';
-import { vAdd } from '../../util/vector.mts';
-import {
-  rcCell,
-  rcDirectedRxBuffer,
-  rcDirectedTxBuffer,
-} from './recoil.app.mts';
-import type { DRCellId, DRDirection } from './util.mts';
-import {
-  DRAdjacentRxDirection,
-  DRAdjacentStep,
-  getAdjacentId,
-} from './util.mts';
-
-export const rcTransmitMessage = writer<{
-  cellId: DRCellId;
-  d: DRDirection;
-}>({
-  key: 'TransmitMessage',
-  set: ({ get, set }, data) => {
-    const { cellId, d } = data;
-    const tx = rcDirectedTxBuffer(`${cellId},${d}`);
-    const buf = get(tx).slice();
-    const tMsg = buf.shift();
-    set(tx, buf);
-    if (!tMsg) {
-      return;
-    }
-    const aId = getAdjacentId(cellId, d);
-    if (!get(rcCell(aId))) {
-      return;
-    }
-    const ad = DRAdjacentRxDirection[d];
-    const rMsg = { ...tMsg, d: vAdd(tMsg.d, DRAdjacentStep[d]) };
-    if (rMsg.ttl) {
-      rMsg.ttl -= 1;
-    }
-    set(rcDirectedRxBuffer(`${aId},${ad}`), (buffer) => [...buffer, rMsg]);
-  },
-});
